Diabetic_Retinopathy_Project/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
albumentations>=1.3.0
opencv-python>=4.7.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pydicom>=2.4.0
Pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.65.0
wandb>=0.15.0
plotly>=5.15.0
jupyter>=1.0.0
ipykernel>=6.25.0
grad-cam>=1.4.8
